Move DestroyedForRecreation to *.Lifecycle
