+++ HoneyComb.Platform.Android/Lifecycle/IsBeingDestroyedForRecreation.cs
@@ -1,7 +1,7 @@
 ﻿using Android.App;
 using Android.OS;
 
-namespace HoneyComb.Platform.Android.OS
+namespace HoneyComb.Platform.Android.Lifecycle
 {
     /// <summary>
     ///     Utility to help standardize the use case of detecting
